[Refactor] #124 - segmented control 뷰 분리
