--- Diptych/Diptych/Presentation/Archive/View/QuestionList/LikeSegmentedControlView.swift
@@ -1,21 +1,0 @@
-//
-//  LikeSegmentedControlView.swift
-//  Diptych
-//
-//  Created by 김민 on 2023/10/08.
-//
-
-import SwiftUI
-
-struct LikeSegmentedControlView: View {
-
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-struct LikeSegmentedControlView_Previews: PreviewProvider {
-    static var previews: some View {
-        LikeSegmentedControlView()
-    }
-}
